+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.0.2](https://github.com/bitrix24/b24ui/compare/v2.0.1...v2.0.2) (2025-10-23)
+
+* chore(deps): update dependency reka-ui to v2.6.0
+
 ## [2.0.1](https://github.com/bitrix24/b24ui/compare/v2.0.0...v2.0.1) (2025-10-22)
 
 ### Features